seed(1428943783): Add v1.4.2 stats
Full log below:

1. Pallet Rival: 0:01:59 (01:59)

is waiting in the grass for a suitable encounter for experience || 202

2. Nidoran grass: 0:05:47 (03:48)

12 11 12 15

beat Brock with a great Nidoran! Att: 16, Def: 13, Speed: 15, Special: 13, caught at level 3.

3. Brock's Gym: 0:10:32 (04:45)

4. Route 3: 0:17:34 (07:02)

10 Moon encounters, but we caught a Paras!

5. Mt. Moon: 0:25:27 (07:53)

RCE 5374

12 11 12 15

55 47 52 48

6. Nugget Bridge: 0:32:42 (07:15)

7. Misty's Gym: 0:38:29 (05:47)

Ugh, 14 try Trashcans || 0:46:47

8. Surge's Gym: 0:47:47 (09:18)

needs a 4-turn thrash (1 in 2 chance) to beat this dangerous trainer...

Giovanni skip strats!

9. HM02 Fly: 0:58:05 (10:18)

is using Rock Slide to one-hit these Ghastlies in red-bar (each is 1 in 10 to miss)

10. Pokemon Tower: 1:05:48 (07:43)

11. Silph Giovanni: 1:16:27 (10:39)

EQ Elixer: Koga's Gym

12. Koga's Gym: 1:19:56 (03:29)

13. Erika's Gym: 1:25:04 (05:08)

14. Blaine's Gym: 1:27:58 (02:54)

15. Sabrina's Gym: 1:29:40 (01:42)

EQ Elixer: Sabrina's Gym

16. Giovanni's Gym: 1:33:15 (03:35)

is skipping the Center and attempting to red-bar the Elite 4!

17. Lorelei: 1:41:07 (07:52)

18. Bruno: 1:42:21 (01:14)

19. Agatha: 1:44:11 (01:50)

20. Lance: 1:46:05 (01:54)

21. Blue: 1:48:55 (02:50)

v1.4.2: 392144 frames, with seed 1428943783

beat the game in 1:48:55!