seed(1022054): Add v1.4.3 stats
Full log below:

1. Pallet Rival: 0:01:56 (01:56)

2. Nidoran grass: 0:06:14 (04:18)

14 10 14 4

beat Brock with a good Nidoran! Att: 16, Def: 13, Speed: 15, Special: 12, caught at level 3.

3. Brock's Gym: 0:10:38 (04:24)

4. Route 3: 0:18:20 (07:42)

8 Moon encounters, but we didn't catch a Paras :(

5. Mt. Moon: 0:25:29 (07:09)

14 10 14 4

56 47 53 43

6. Nugget Bridge: 0:32:51 (07:22)

Thrash didn't finish the kill :( swapping to Squirtle for safety

is limiting heals to attempt to get closer to red-bar off Misty... || 5

is limiting heals to attempt to get closer to red-bar off Misty... || 4

7. Misty's Gym: 0:38:32 (05:41)

is searching for an Oddish in the grass, to teach it Cut

Good, 7 try Trashcans! || 0:47:28

8. Surge's Gym: 0:48:36 (10:04)

needs a 4-turn thrash (1 in 2 chance) to beat this dangerous trainer...

is using Thunderbolt to attempt to redbar off Cubone

Giovanni skip strats!

9. HM02 Fly: 0:58:59 (10:23)

is using Rock Slide to one-hit these Ghastlies in red-bar (each is 1 in 10 to miss)

10. Pokemon Tower: 1:06:44 (07:45)

gen 1 missed :( (1 in 256 chance)

11. Silph Giovanni: 1:17:37 (10:53)

EQ Elixer: Koga's Gym

12. Koga's Gym: 1:21:15 (03:38)

13. Erika's Gym: 1:26:30 (05:15)

14. Blaine's Gym: 1:29:25 (02:55)

EQ Elixer: Saffron City

15. Sabrina's Gym: 1:31:11 (01:46)

is using risky strats on Giovanni to skip the extra Max Ether...

16. Giovanni's Gym: 1:34:36 (03:25)

is skipping the Center and attempting to red-bar the Elite 4!

Swapping out Pidgey to tank Aurora Beam into turn 2 Rest. Only a problem if it misses...

17. Lorelei: 1:42:28 (07:52)

18. Bruno: 1:43:49 (01:21)

19. Agatha: 1:45:39 (01:50)

20. Lance: 1:47:34 (01:55)

21. Blue: 1:50:37 (03:03)

v1.4.3: 398241 frames, with seed 1022054

beat the game in 1:50:37!